Jest: Add list plugin action tests
